Split peak binning modules
