+++ apps/server/tests/use_cases/diagnostics/test_finding_objects.py
@@ -1,13 +1,11 @@
-"""Tests for finalize_findings, PeakBin, and PeakFindingAnalyzer."""
+"""Tests for finalize_findings and PeakFindingAnalyzer."""
 
 from __future__ import annotations
 
-import pytest
 from test_support.findings import make_finding_payload, make_info_finding, make_ref_finding
 
 from vibesensor.shared.boundaries.finding import finding_from_payload
 from vibesensor.use_cases.diagnostics.findings import PeakFindingAnalyzer, finalize_findings
-from vibesensor.use_cases.diagnostics.peak_binning import PeakBin
 
 # ===========================================================================
 # Domain Finding from payload (replaces former FindingRecord tests)
@@ -99,140 +97,6 @@ class TestFinalizeFindings:
     def test_finalize_empty(self) -> None:
         domain_findings = finalize_findings([])
         assert domain_findings == ()
-
-
-# ===========================================================================
-# PeakBin
-# ===========================================================================
-
-
-def _make_peak_bin(
-    *,
-    bin_center: float = 50.0,
-    amps: list[float] | None = None,
-    n_samples: int = 100,
-    total_locations: set[str] | None = None,
-    run_noise_baseline_g: float | None = 0.001,
-) -> PeakBin:
-    """Helper to build a PeakBin with sensible defaults."""
-    if amps is None:
-        amps = [0.05] * 50  # Moderate amplitude, 50% presence
-    if total_locations is None:
-        total_locations = {"front_left"}
-    n = len(amps)
-    return PeakBin(
-        bin_center=bin_center,
-        amps=amps,
-        floor_vals=[0.005] * n,
-        speed_amp_pairs=[(60.0, a) for a in amps],
-        loc_counts_for_bin={"front_left": n},
-        speed_bin_counts_for_bin={"60-80": n},
-        phases_for_bin={},
-        n_samples=n_samples,
-        total_locations=total_locations,
-        total_location_sample_counts={"front_left": n_samples},
-        total_speed_bin_counts={"60-80": n_samples},
-        run_noise_baseline_g=run_noise_baseline_g,
-        has_phases=False,
-    )
-
-
-class TestPeakBin:
-    def test_bin_center(self) -> None:
-        pb = _make_peak_bin(bin_center=42.0)
-        assert pb.bin_center == 42.0
-
-    def test_presence_ratio(self) -> None:
-        # 50 amps out of 100 samples = 0.50
-        pb = _make_peak_bin(amps=[0.05] * 50, n_samples=100)
-        assert pb.presence_ratio == pytest.approx(0.50)
-
-    def test_burstiness_uniform(self) -> None:
-        # All same amplitude → burstiness = 1.0
-        pb = _make_peak_bin(amps=[0.05] * 20)
-        assert pb.burstiness == pytest.approx(1.0)
-
-    def test_burstiness_spiky(self) -> None:
-        # One spike, rest low → high burstiness
-        amps = [0.01] * 19 + [1.0]
-        pb = _make_peak_bin(amps=amps)
-        assert pb.burstiness > 5.0
-
-    def test_snr_positive(self) -> None:
-        pb = _make_peak_bin()
-        assert pb.snr > 0.0
-
-    def test_peak_type_patterned(self) -> None:
-        # High presence, low burstiness, good SNR → patterned
-        pb = _make_peak_bin(amps=[0.05] * 80, n_samples=100)
-        assert pb.peak_type == "patterned"
-
-    def test_peak_type_transient(self) -> None:
-        # Very low presence → transient
-        pb = _make_peak_bin(amps=[0.05] * 5, n_samples=100)
-        assert pb.peak_type == "transient"
-
-    def test_is_transient_property(self) -> None:
-        pb = _make_peak_bin(amps=[0.05] * 5, n_samples=100)
-        assert pb.is_transient is True
-
-    def test_confidence_bounded(self) -> None:
-        pb = _make_peak_bin()
-        assert 0.0 <= pb.confidence <= 1.0
-
-    def test_ranking_score_positive(self) -> None:
-        pb = _make_peak_bin()
-        assert pb.ranking_score > 0.0
-
-    def test_to_finding_has_required_keys(self) -> None:
-        pb = _make_peak_bin()
-        finding = pb.to_finding()
-        assert finding.finding_id == "F_PEAK"
-        assert str(finding.suspected_source)
-        assert finding.confidence is not None
-        assert finding.kind is not None
-
-    def test_to_finding_preserves_bin_center(self) -> None:
-        pb = _make_peak_bin(bin_center=42.0)
-        finding = pb.to_finding()
-        assert finding.order == "42.0 Hz"
-
-    def test_to_finding_includes_evidence_metrics(self) -> None:
-        pb = _make_peak_bin()
-        finding = pb.to_finding()
-        metrics = finding.evidence
-        assert metrics is not None
-        assert metrics.presence_ratio > 0.0
-        assert metrics.burstiness >= 0.0
-        assert metrics.spatial_concentration > 0.0
-
-    def test_spatial_uniformity_single_location(self) -> None:
-        pb = _make_peak_bin(total_locations={"front_left"})
-        assert pb.spatial_uniformity is None
-
-    def test_spatial_uniformity_multi_location(self) -> None:
-        pb = PeakBin(
-            bin_center=50.0,
-            amps=[0.05] * 40,
-            floor_vals=[0.005] * 40,
-            speed_amp_pairs=[(60.0, 0.05)] * 40,
-            loc_counts_for_bin={"front_left": 20, "front_right": 20},
-            speed_bin_counts_for_bin={"60-80": 40},
-            phases_for_bin={},
-            n_samples=100,
-            total_locations={"front_left", "front_right", "rear_left", "rear_right"},
-            total_location_sample_counts={
-                "front_left": 25,
-                "front_right": 25,
-                "rear_left": 25,
-                "rear_right": 25,
-            },
-            total_speed_bin_counts={"60-80": 100},
-            run_noise_baseline_g=0.001,
-            has_phases=False,
-        )
-        assert pb.spatial_uniformity is not None
-        assert pb.spatial_uniformity == pytest.approx(0.50)
 
 
 # ===========================================================================
