+++ apps/server/vibesensor/use_cases/diagnostics/peak_classification.py
@@ -1,0 +1,66 @@
+"""Peak classification thresholds and policy for diagnostics."""
+
+from __future__ import annotations
+
+PERSISTENT_PEAK_MIN_PRESENCE = 0.15
+TRANSIENT_BURSTINESS_THRESHOLD = 5.0
+
+# Minimum SNR for a peak to be considered above baseline noise.
+BASELINE_NOISE_SNR_THRESHOLD = 1.5
+
+# High spatial uniformity: present across most sensor locations -> likely noise.
+_SPATIAL_UNIFORMITY_HIGH = 0.85
+# Medium spatial uniformity: used with speed-uniformity check.
+_SPATIAL_UNIFORMITY_MED = 0.80
+# Presence ratio below which a "high spatial uniformity" peak is noise.
+_NOISE_PRESENCE_MIN_HIGH = 0.60
+# Burstiness ceiling for "spatially uniform + high presence" noise check.
+_NOISE_BURSTINESS_MAX_LOW = 2.0
+# Speed-uniformity (std-dev) ceiling: flat across speed bins -> noise.
+_NOISE_SPEED_UNIFORMITY_MAX = 0.10
+# Presence band for the "medium spatial + low speed variance" noise check.
+_NOISE_PRESENCE_LOW_MIN = 0.20
+_NOISE_PRESENCE_LOW_MAX = 0.40
+# Burstiness band for the "medium spatial + low speed variance" noise check.
+_NOISE_BURSTINESS_BAND_MIN = 3.0
+_NOISE_BURSTINESS_BAND_MAX = 5.0
+# Minimum presence and maximum burstiness for a "patterned" peak.
+_PATTERNED_MIN_PRESENCE = 0.40
+_PATTERNED_MAX_BURSTINESS = 3.0
+
+
+def classify_peak_type(
+    presence_ratio: float,
+    burstiness: float,
+    *,
+    snr: float | None = None,
+    spatial_uniformity: float | None = None,
+    speed_uniformity: float | None = None,
+) -> str:
+    """Classify a frequency peak as patterned/persistent/transient/baseline noise."""
+    if snr is not None and snr < BASELINE_NOISE_SNR_THRESHOLD:
+        return "baseline_noise"
+    if (
+        spatial_uniformity is not None
+        and spatial_uniformity > _SPATIAL_UNIFORMITY_HIGH
+        and presence_ratio >= _NOISE_PRESENCE_MIN_HIGH
+        and burstiness < _NOISE_BURSTINESS_MAX_LOW
+    ):
+        return "baseline_noise"
+    if (
+        spatial_uniformity is not None
+        and speed_uniformity is not None
+        and spatial_uniformity >= _SPATIAL_UNIFORMITY_MED
+        and speed_uniformity <= _NOISE_SPEED_UNIFORMITY_MAX
+        and _NOISE_PRESENCE_LOW_MIN <= presence_ratio <= _NOISE_PRESENCE_LOW_MAX
+        and _NOISE_BURSTINESS_BAND_MIN <= burstiness <= _NOISE_BURSTINESS_BAND_MAX
+    ):
+        return "baseline_noise"
+
+    if presence_ratio < PERSISTENT_PEAK_MIN_PRESENCE:
+        return "transient"
+    if burstiness > TRANSIENT_BURSTINESS_THRESHOLD:
+        return "transient"
+    if presence_ratio >= _PATTERNED_MIN_PRESENCE and burstiness < _PATTERNED_MAX_BURSTINESS:
+        return "patterned"
+    return "persistent"
