+++ docs/report_pipeline.md
@@ -21,7 +21,7 @@ Recording stops
       → RunAnalysis(...).summarize() [vibesensor.use_cases.diagnostics.summary_builder]
       → run_data_preparation.py + _summary_steps.py + _summary_result.py + summary_builder.py (preparation, phases, suitability, domain/result assembly)
       → analysis_result_to_summary() [vibesensor.shared.boundaries.analysis_summary]
-      → findings.py + _peak_findings.py + _reference_findings.py, peak_binning.py, signal_aggregation.py, top_cause_selection.py, plots.py, peak_table.py
+      → findings.py + _peak_findings.py + _reference_findings.py, peak_accumulation.py, peak_classification.py, peak_scoring.py, peak_finding_builder.py, signal_aggregation.py, top_cause_selection.py, plots.py, peak_table.py
     → map_summary() [vibesensor.adapters.pdf.mapping]
       → report_context.py (context assembly, card decisions) + mapping.py (thin template mapper) + peak_table.py + report_sections.py
     → store_analysis() [vibesensor.adapters.persistence.history_db]
